+++ src/Altinn.App.Core/Implementation/ProcessChangeHandler.cs
@@ -5,7 +5,6 @@ using Altinn.App.Common.Process.Elements;
 using Altinn.App.Core.Interface;
 using Altinn.App.Core.Models;
 using Altinn.App.Core.Process;
-using Altinn.App.Core.Receivers;
 using Altinn.App.PlatformServices.Extensions;
 using Altinn.App.PlatformServices.Interface;
 using Altinn.App.Services.Configuration;
